Move Objective-C interop code to discrete source files
