+++ Sources/ObjC/PusherError+ObjectiveC.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+public extension PusherError {
+    /// The error code as an NSNumber (for Objective-C compatibility).
+    var codeOC: NSNumber? {
+        guard let code = code else {
+            return nil
+        }
+
+        return NSNumber(value: code)
+    }
+}
